fixed function for spawning squig model and replaced rat with squig
Function for spawning squig model can cause crashing when attempting to walk through the squig model will probably not fix this crash. Updated the state machine and animations to allow the critter rat to be replaced with a squig model. Also reemoved the physx from the model and replaced it with death animations to let the running animations work.